+++ PierreBakeryVendors/Controllers/VendorsController.cs
@@ -40,7 +40,7 @@ namespace PierreBakeryVendors.Controllers
     }
     
     [HttpPost("/vendors/{vendorId}/orders")]
-    public ActionResult Create(int vendorId, string orderTitle, string orderDescription, string orderDate, int orderPrice)
+    public ActionResult Create(int vendorId, string orderTitle, string orderDescription, string orderDate, float orderPrice)
     {
       Dictionary<string, object> model = new Dictionary<string, object>();
       Vendor foundVendor = Vendor.Find(vendorId);
